Implemented cucumber test. Migrated to BigDecimal.

+++ src/main/java/com/test/scraper/bean/TotalBean.java
@@ -2,12 +2,14 @@ package com.test.scraper.bean;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @EqualsAndHashCode
 @NoArgsConstructor
 @AllArgsConstructor
 public class TotalBean {
-    private Double gross;
-    private Double vat;
+    private BigDecimal gross;
+    private BigDecimal vat;
 }
